Make neo_zoom() return boolean (#2)
Without this, it's not possible to know if the toggle was skipped due to e.g. the buffer being a terminal or the filetype being in the exclude list. This makes it difficult and awkward to code around those situations in custom configuration. By making it a boolean, we can call the function directly (rather than using vim cmd) and use the result to know if it was skipped or not.